+++ realm/realm-library/src/main/java/io/realm/internal/OsResults.java
@@ -322,7 +322,11 @@ public class OsResults implements NativeObject, ObservableCollection {
 
     public OsResults freeze() {
         OsSharedRealm frozenRealm = sharedRealm.freeze();
-        return new OsResults(frozenRealm, table.freeze(), nativeFreeze(nativePtr, frozenRealm.getNativePtr()));
+        OsResults results = new OsResults(frozenRealm, table.freeze(), nativeFreeze(nativePtr, frozenRealm.getNativePtr()));
+        if (isLoaded()) {
+            results.load();
+        }
+        return results;
     }
 
     @Override